feat(GAT-7861): Use deployed MJML (#1472)

+++ database/migrations/2024_01_04_110820_replacement-fields-in-email-templates.php
@@ -20,8 +20,10 @@ return new class () extends Migration {
      */
     public function down(): void
     {
-        Schema::table('email_templates', function (Blueprint $table) {
-            $table->dropColumn('buttons');
-        });
+        if (Schema::hasColumn('email_templates', 'buttons')) {
+            Schema::table('email_templates', function (Blueprint $table) {
+                $table->dropColumn('buttons');
+            });
+        }
     }
 };
